// src/app/interfaces/sodetails.interface.ts
export interface so_detailsInterface {
  order_id: string;
  productid: string;
  productname: string;
  imageurl: string;
  order_line_no: number;
  product_qty: number;
  product_mrp: number;
  product_unit_price_applied: never;
  product_coupon_code_applied: string;
  product_discount_code_applied: string;
  updated_at: string;
  update_by: string;
  customer_id: string;
}